requirements.txt
aniso8601==7.0.0
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
blinker==1.4
bson==0.5.8
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
coverage==4.5.3
Flask==1.0.3
Flask-API==1.0
Flask-Cors==3.0.8
flask-mongoengine==0.9.5
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-WTF==0.14.2
gunicorn==19.9.0
idna==2.8
importlib-metadata==0.18
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
marshmallow==3.0.0rc7
marshmallow-mongoengine==0.9.1
mccabe==0.6.1
mongoengine==0.18.2
more-itertools==7.0.0
packaging==19.0
pluggy==0.12.0
py==1.8.0
pycparser==2.19
PyJWT==1.7.1
pylint==2.3.1
pymongo==3.8.0
pyparsing==2.4.0
pytest==4.6.3
pytest-cov==2.7.1
python-dateutil==2.8.0
python-decouple==3.1
python-http-client==3.1.0
pytz==2019.1
requests==2.22.0
six==1.12.0
slack==0.0.2
slackclient==2.1.0
typed-ast==1.4.0
tzlocal==1.5.1
urllib3==1.25.3
wcwidth==0.1.7
websocket-client==0.56.0
Werkzeug==0.15.4
wrapt==1.11.2
WTForms==2.2.1
zipp==0.5.1
